# socialapp/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: ignacio.gomez.arboleda@gmail.com
    name: Socialapp
    url: https://socialapp.gomezignacio.com
  description: Socialapp is a generic social network.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Socialapp
  version: 1.0.0
servers:
  - url: https://socialapp.gomezignacio.com
  - url: http://localhost:8080
  - url: http://localhost:8085
...
security:
  - OAuth2: []
paths:
...
  /v1/oauth/token:
    post:
      description: Get an access token
      operationId: getAccessToken
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessToken"
          description: Access token
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      security:
        - BasicAuth: []
      summary: Get an access token
      tags:
        - Authentication
...
components:
  schemas:
...
    AccessToken:
      example:
        access_token: 123token123
        token_type: Bearer
        expires_in: 3600
      properties:
        access_token:
          type: string
        token_type:
          type: string
        scopes:
          items:
            type: string
          type: array
        expires_in:
          type: integer
      required:
        - access_token
        - expires_in
        - token_type
      type: object
    Error:
      example:
        code: 404
        message: Not found
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
        - code
        - message
      type: object
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    OAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /v1/oauth/token
          scopes:
            socialapp.users.list: List users
            socialapp.users.create: Create users
            socialapp.users.update: Update users
            socialapp.users.delete: Delete users
            socialapp.users.read: Read a user
            socialapp.comments.list: List comments
            socialapp.comments.create: Create comments
            socialapp.comments.update: Update comments
            socialapp.comments.delete: Delete comments
            socialapp.followers.list: List followers
            socialapp.following.list: List following
            socialapp.roles.list: List roles
            socialapp.roles.create: Create roles
            socialapp.roles.read: Read a role
            socialapp.roles.update: Update roles
            socialapp.roles.delete: Delete roles
            socialapp.scopes.list: List scopes
            socialapp.scopes.create: Create scopes
            socialapp.scopes.read: Read a scope
            socialapp.scopes.update: Update scopes
            socialapp.scopes.delete: Delete scopes
            socialapp.roles.list_scopes: List scopes of a role
            socialapp.roles.scopes.create: Create scopes of a role
            socialapp.roles.scopes.delete: Delete scopes of a role
            socialapp.users.roles.list: List roles of a user
            socialapp.users.roles.create: Create roles of a user
            socialapp.users.roles.update: Update roles of a user
            socialapp.users.roles.delete: Delete roles of a user
            shortly.url.create: Create a url
            shortly.url.update: Update a url
            shortly.url.delete: Delete a url